Convert alignment icons to SVG and create 100% and 200% zoom variants
Used by e.g. the MigLayout in Swing as well as in SWT, in order to
specify the alignment of the fill behavior. In order to reduce the
amount of redundancy, common icons should be stored in a common base
plugin, rather than each plugin keeping a copy of its used images.